requirements.txt
aiogram==3.4.1
sqlalchemy==2.0.25
aiosqlite==0.19.0
selenium==4.18.1
python-dotenv==1.0.0
